Potential fix for circular imports
