+++ src/haplovis/server/managers/gfa.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pathlib import Path
 from typing import Dict, List, Optional, Set
 from haplovis.schemas.gfa import Gfa as GfaDataclass, GfaLink, GfaPath, GfaSegment
@@ -135,7 +137,7 @@ class GfaManager:
     def preprocess(cls) -> None:
         cls.prepare_gfa()
         file_path = FileManager.get_absolute_file_path(FileIndex.GFA)
-        layout = Layout.compute_layout(file_path)
+        layout = managers.LayoutManager.compute_layout(file_path)
         LayoutManager.index = LayoutManager.get_index_from_layout(layout)
         gfa_hash = Gfa.get_gfa_hash(file_path)
         if gfa_hash:
